% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_elicitation.R
\name{get_all_elicitations}
\alias{get_all_elicitations}
\title{Computes \code{mu_0}, \code{lambda_0}, \code{alpha_0}, \code{beta_0} using elicitation for a all
features. This function corresponds to part (i) of the algorithm in the
paper.}
\usage{
get_all_elicitations(features, initialization_function = NULL)
}
\arguments{
\item{features}{(list)

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(sample1 = c(1, 2, 3), sample2 = c(4, 5, 6), ...)
}\if{html}{\out{</div>}}}

\item{initialization_function}{(function) should return \code{list(a, b, s_1, s_2)} Default: \code{get_single_a_b_s1_s2()}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  sample1 = list(mu_0, lambda_0, alpha_0, beta_0),
  sample2 = ...
)
}\if{html}{\out{</div>}}
}
\description{
Computes \code{mu_0}, \code{lambda_0}, \code{alpha_0}, \code{beta_0} using elicitation for a all
features. This function corresponds to part (i) of the algorithm in the
paper.
}
