% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma_sq_and_mu.R
\name{get_all_sigma_sq_and_mu}
\alias{get_all_sigma_sq_and_mu}
\title{Computes prior & posterior \code{sigma_sq} and \code{mu} for all features.
This function corresponds to parts (ii) & (v) of the algorithm in the paper.}
\usage{
get_all_sigma_sq_and_mu(features, initialization_function = NULL)
}
\arguments{
\item{features}{(list)}

\item{initialization_function}{(function) returns \code{list(a, b, s_1, s_2)}.
Default: \code{get_single_a_b_s1_s2()}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  sample1 = list(
    sigma_sq = list(prior, post),
    mu = list(prior, post)
  ),
  sample2 = ...
)
}\if{html}{\out{</div>}}
}
\description{
Computes prior & posterior \code{sigma_sq} and \code{mu} for all features.
This function corresponds to parts (ii) & (v) of the algorithm in the paper.
}
