% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_divergence.R
\name{get_divergence}
\alias{get_divergence}
\title{Computes the prior & posterior KL divergence values.
This function corresponds to parts (iii) & (vi) of the algorithm in the paper}
\usage{
get_divergence(features, initialization_function = NULL)
}
\arguments{
\item{features}{(list)}

\item{initialization_function}{(function) returns \code{list(a, b, s_1, s_2)}.
Default: \code{get_single_a_b_s1_s2()}.}
}
\value{
\code{list(prior, post)}
}
\description{
Computes the prior & posterior KL divergence values.
This function corresponds to parts (iii) & (vi) of the algorithm in the paper
}
