% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_divergence.R
\name{get_divergence_}
\alias{get_divergence_}
\title{Computes the prior & posterior KL divergence values.}
\usage{
get_divergence_(features, mu_0_lambda_0_alpha_0_beta_0)
}
\arguments{
\item{features}{(list)

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(sample1 = c(1, 2, 3), sample2 = c(4, 5, 6), ...)
}\if{html}{\out{</div>}}}

\item{mu_0_lambda_0_alpha_0_beta_0}{(list)

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  sample1 = list(mu_0, lambda_0, alpha_0, beta_0),
  sample2 = ...
)
}\if{html}{\out{</div>}}}
}
\value{
\code{c(prior, post)}
}
\description{
Computes the prior & posterior KL divergence values.
}
