% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_divergence.R
\name{get_divergence_distribution}
\alias{get_divergence_distribution}
\title{Computes a distribution of the prior & posterior KL divergence values.
This function corresponds to parts (iii) & (vi) of the algorithm in the paper}
\usage{
get_divergence_distribution(
  features,
  initialization_function = NULL,
  repetition = 5000
)
}
\arguments{
\item{features}{(list)

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(sample1 = c(1, 2, 3), sample2 = c(4, 5, 6), ...)
}\if{html}{\out{</div>}}}

\item{initialization_function}{(function)
should return \code{list(a, b, s_1, s_2)}
Default: \code{get_single_a_b_s1_s2()}.}

\item{repetition}{(integer)
number of times to repeat the computation
(i.e.: number of samples in the distribution)}
}
\value{
matrix of size \code{length(features)} x \code{repetition}
each row corresponds to a feature
}
\description{
Computes a distribution of the prior & posterior KL divergence values.
This function corresponds to parts (iii) & (vi) of the algorithm in the paper
}
