% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_a_b_s1_s2.R
\name{get_single_a_b_s1_s2}
\alias{get_single_a_b_s1_s2}
\title{Calculate the confidence intervals of the mean and
standard deviation for a given feature.
This function corresponds to the part "before" part (i) of
the algorithm in the paper.}
\usage{
get_single_a_b_s1_s2(feature, conf_level = 0.95)
}
\arguments{
\item{feature}{(vector)}

\item{conf_level}{(number) should be between 0 and 1, both exclusive}
}
\value{
\code{list(a, b, s_1, s_2)} where a and b are the lower and upper
bound of the mean and s_1 and s_2 are the lower and upper bound
of the standard deviation.
}
\description{
Calculate the confidence intervals of the mean and
standard deviation for a given feature.
This function corresponds to the part "before" part (i) of
the algorithm in the paper.
}
