% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_sigma_sq_and_mu.R
\name{get_single_sigma_sq_and_mu}
\alias{get_single_sigma_sq_and_mu}
\title{Computes \code{sigma_sq} and \code{mu} for a single feature.
This function corresponds to part (ii) of the algorithm in the paper.}
\usage{
get_single_sigma_sq_and_mu(mu_0, lambda_0, alpha_0, beta_0)
}
\arguments{
\item{mu_0}{(number)}

\item{lambda_0}{(number)}

\item{alpha_0}{(number)}

\item{beta_0}{(number)}
}
\value{
\code{list(sigma_sq, mu)}
}
\description{
Computes \code{sigma_sq} and \code{mu} for a single feature.
This function corresponds to part (ii) of the algorithm in the paper.
}
