% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma_sq_and_mu.R
\name{get_single_sigma_sq_mu_prior_}
\alias{get_single_sigma_sq_mu_prior_}
\title{Computes \strong{\emph{prior}} \code{sigma_sq} and \code{mu} for a single feature.}
\usage{
get_single_sigma_sq_mu_prior_(mu_0, lambda_0, alpha_0, beta_0)
}
\arguments{
\item{mu_0}{(number)}

\item{lambda_0}{(number)}

\item{alpha_0}{(number)}

\item{beta_0}{(number)}
}
\value{
\code{list(sigma_sq, mu)}
}
\description{
Computes \strong{\emph{prior}} \code{sigma_sq} and \code{mu} for a single feature.
}
