% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relbel-package.R
\docType{package}
\name{relbel-package}
\alias{relbel-package}
\title{relbel: Computes the Relative Belief Ratio, a k-sample Bayesian Test for Mean}
\description{
Computes the Relative Belief Ratio, a k-sample Bayesian Test for Mean.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ahmad-Alsaleh/relbel}
  \item \url{https://ahmad-alsaleh.github.io/relbel/}
  \item Report bugs at \url{https://github.com/Ahmad-Alsaleh/relbel/issues}
}

}
\author{
\strong{Maintainer}: Ahmad Alsaleh \email{ahmed.asaleh2@gmail.com}

}
\keyword{internal}
