% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relbel.R
\name{relbel}
\alias{relbel}
\title{Compute Relative Belief Ratio (RB) Score.}
\usage{
relbel(
  x,
  y,
  initialization_function = NULL,
  repetition = 5000,
  L = 20,
  i_0 = 1
)
}
\arguments{
\item{x}{a numeric vector or matrix; the feature(s).}

\item{y}{a}

\item{initialization_function}{A function used for initializating the main and standard deviation. See 'Details' for more information.}

\item{repetition}{The number of repetitions.}

\item{L}{The value of L.}

\item{i_0}{The value of i_0.}
}
\value{
The computed relative belief ratio.
}
\description{
This function computes the relative belief ratio based on the given features
and parameters.
}
\examples{
initialization_function <- function(feature) {
  list(a = 0, b = 2, s_1 = 2, s_2 = 4)
}
repetition <- 5000
L <- 20
i_0 <- 1

dataset <- data.frame(
  x1 = c(-0.8606568, 1.4141698, 0.2235239, -1.1968656, 0.7677741),
  x2 = c(-0.318, -0.712, -1.320, -0.008, 0.048),
  y = c("A", "B", "A", "B", "B")
)

relbel(
  dataset[1:2], dataset["y"],
  initialization_function, repetition, L, i_0
)

}
\references{
TODO:
}
\author{
Ahmad Alsaleh
}
