% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_sigma_sq_and_mu.R
\name{get_all_sigma_sq_and_mu}
\alias{get_all_sigma_sq_and_mu}
\title{Computes prior & posterior \code{sigma_sq} and \code{mu} for all features. This
function corresponds to parts (ii) & (v) of the algorithm in the paper.}
\usage{
get_all_sigma_sq_and_mu(features, mu_0_lambda_0_alpha_0_beta_0)
}
\arguments{
\item{features}{(list)

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(sample1 = c(1, 2, 3), sample2 = c(4, 5, 6), ...)
}\if{html}{\out{</div>}}}

\item{mu_0_lambda_0_alpha_0_beta_0}{(list)

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  sample1 = list(mu_0, lambda_0, alpha_0, beta_0),
  sample2 = ...
)
}\if{html}{\out{</div>}}}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  sample1 = list(
    sigma_sq = list(prior, post),
    mu = list(prior, post)
  ),
  sample2 = ...
)
}\if{html}{\out{</div>}}
}
\description{
Computes prior & posterior \code{sigma_sq} and \code{mu} for all features. This
function corresponds to parts (ii) & (v) of the algorithm in the paper.
}
