% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mu_star.R
\name{get_mu_star}
\alias{get_mu_star}
\title{Computes prior & posterior \code{mu_star}.}
\usage{
get_mu_star(features, sigma_sq_and_mu)
}
\arguments{
\item{features}{(list)}

\item{sigma_sq_and_mu}{(list with the following structure)

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  sample1 = list(
    sigma_sq = list(prior, post),
    mu = list(prior, post)
  ),
  sample2 = ...
)
}\if{html}{\out{</div>}}}
}
\value{
\code{list(prior, post)}
}
\description{
Computes prior & posterior \code{mu_star}.
}
