% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mu_star.R
\name{get_mu_star}
\alias{get_mu_star}
\title{Computes \code{mu_star} # ! TODO: write this.}
\usage{
get_mu_star(features, sigma_sq_and_mu)
}
\arguments{
\item{features}{foo}

\item{sigma_sq_and_mu}{boo}
}
\value{
bar
}
\description{
Computes \code{mu_star} # ! TODO: write this.
}
